+++ lib/metronome_platform_interface.dart
@@ -37,7 +37,7 @@ abstract class MetronomePlatform extends PlatformInterface {
     int bpm = 120,
     int volume = 50,
     bool enableTickCallback = false,
-    int timeSignature = 4,
+    List<int> accentPattern = const [4],
     int sampleRate = 44100,
     bool manageAudioSession = true,
   }) {
@@ -83,6 +83,10 @@ abstract class MetronomePlatform extends PlatformInterface {
     throw UnimplementedError('getBPM() has not been implemented.');
   }
 
+  Future<void> setAccentPattern(List<int> accentPattern) {
+    throw UnimplementedError('setAccentPattern() has not been implemented.');
+  }
+
   Future<void> configureTempoRamp(TempoRampConfig config) {
     throw UnimplementedError(
       'configureTempoRamp() has not been implemented.',
@@ -93,12 +97,8 @@ abstract class MetronomePlatform extends PlatformInterface {
     throw UnimplementedError('disableTempoRamp() has not been implemented.');
   }
 
-  Future<void> setTimeSignature(int timeSignature) {
-    throw UnimplementedError('setTimeSignature() has not been implemented.');
-  }
-
-  Future<int?> getTimeSignature() {
-    throw UnimplementedError('getTimeSignature() has not been implemented.');
+  Future<List<int>?> getAccentPattern() {
+    throw UnimplementedError('getAccentPattern() has not been implemented.');
   }
 
   Future<void> destroy() {