+++ lib/metronome_web.dart
@@ -29,7 +29,7 @@ class MetronomeWeb extends MetronomePlatform {
   bool _isPlaying = false;
   int _currentTick = 0;
   int _bpm = 120;
-  int _timeSignature = 4;
+  List<int> _accentPattern = const [4];
   double _volume = 1.0;
   bool _enableTickCallback = false;
   int _sampleRate = 44100;
@@ -43,6 +43,28 @@ class MetronomeWeb extends MetronomePlatform {
   int _rampStageIndex = 0;
   int _completedRampMeasures = 0;
 
+  int get _totalBeats => _accentPattern.fold<int>(0, (sum, g) => sum + g);
+
+  bool _isAccentBeat(int tick) {
+    int pos = 0;
+    for (final group in _accentPattern) {
+      if (tick == pos) return true;
+      pos += group;
+    }
+    return false;
+  }
+
+  void _validateAccentPattern(List<int> accentPattern) {
+    if (accentPattern.isEmpty) {
+      throw Exception('accentPattern must not be empty');
+    }
+    for (final g in accentPattern) {
+      if (g < 1) {
+        throw Exception('accentPattern groups must be >= 1');
+      }
+    }
+  }
+
   @override
   Future<void> init(
     String mainPath, {
@@ -50,10 +72,11 @@ class MetronomeWeb extends MetronomePlatform {
     int bpm = 120,
     int volume = 50,
     bool enableTickCallback = false,
-    int timeSignature = 4,
+    List<int> accentPattern = const [4],
     int sampleRate = 44100,
     bool manageAudioSession = true,
   }) async {
+    _validateAccentPattern(accentPattern);
     _sampleRate = sampleRate;
     _audioContext = web.AudioContext(
       web.AudioContextOptions(
@@ -69,7 +92,7 @@ class MetronomeWeb extends MetronomePlatform {
       _accentedSoundBuffer = await _bytesToAudioBuffer(accentedPath);
     }
     _bpm = bpm;
-    _timeSignature = timeSignature;
+    _accentPattern = List<int>.from(accentPattern);
     _volume = volume / 100;
     _enableTickCallback = enableTickCallback;
   }
@@ -125,8 +148,8 @@ class MetronomeWeb extends MetronomePlatform {
   }
 
   @override
-  Future<int?> getTimeSignature() async {
-    return _timeSignature;
+  Future<List<int>?> getAccentPattern() async {
+    return List<int>.from(_accentPattern);
   }
 
   @override
@@ -150,6 +173,12 @@ class MetronomeWeb extends MetronomePlatform {
   }
 
   @override
+  Future<void> setAccentPattern(List<int> accentPattern) async {
+    _validateAccentPattern(accentPattern);
+    _accentPattern = List<int>.from(accentPattern);
+  }
+
+  @override
   Future<void> configureTempoRamp(TempoRampConfig config) async {
     config.validate();
     if (_isPlaying) {
@@ -167,13 +196,6 @@ class MetronomeWeb extends MetronomePlatform {
     _rampStageIndex = 0;
     _completedRampMeasures = 0;
     tempoRampController.add(const TempoRampProgress.idle());
-  }
-
-  @override
-  Future<void> setTimeSignature(int timeSignature) async {
-    if (timeSignature != _timeSignature) {
-      _timeSignature = timeSignature;
-    }
   }
 
   @override
@@ -217,8 +239,7 @@ class MetronomeWeb extends MetronomePlatform {
   void _scheduleBeat(double time) {
     _applyRampAtDownbeat();
     final scheduledTick = _currentTick;
-    final effectiveTimeSignature = _timeSignature.clamp(1, 1000);
-    final isAccented = (scheduledTick % effectiveTimeSignature) == 0;
+    final isAccented = _isAccentBeat(scheduledTick);
     final buffer = isAccented ? _accentedSoundBuffer : _mainSoundBuffer;
     final source = _audioContext!.createBufferSource();
     source.buffer = buffer;
@@ -246,7 +267,8 @@ class MetronomeWeb extends MetronomePlatform {
         (delay < 0 ? 0 : delay).toJS,
       );
     }
-    _currentTick = (_currentTick + 1) % effectiveTimeSignature;
+    final totalBeats = _totalBeats;
+    _currentTick = totalBeats < 1 ? 0 : (_currentTick + 1) % totalBeats;
     if (_currentTick == 0 && _rampStatus == TempoRampStatus.running) {
       _completedRampMeasures++;
       _emitRampProgress();